fix(lib): fix textarea multiline autosize problem
