+++ packages/common/src/message-input/message-input.tsx
@@ -141,16 +141,6 @@ export const useMessageInputCore = (props: CommonMessageInputProps) => {
     }
   };
 
-  const handleFileChange = (event: ChangeEvent<HTMLInputElement>) => {
-    try {
-      const file = event.target.files[0];
-      setFile(file);
-      setText(file.name);
-    } catch (e) {
-      onError(e);
-    }
-  };
-
   /*
   /* Lifecycle
   */
@@ -169,7 +159,7 @@ export const useMessageInputCore = (props: CommonMessageInputProps) => {
   return {
     clearInput,
     file,
-    handleFileChange,
+    setFile,
     handleInputChange,
     isValidInputText,
     loader,
